Integrated `conversion-verifier` in CI

Fixed typos


Implemented suggestions


Added initializing echo to `verify-conversions` make target
